# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A PLATFORM OPERATIONS API in golang using Gin framework",
        "title": "PLATFORM OPERATIONS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/argoworkflow": {
            "post": {
                "description": "Create a ArgoWorkflow and saves it in postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new ArgoWorkflow",
                "parameters": [
                    {
                        "description": "Create ArgoWorkflow",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateArgoWorkflowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerArgoWorkflow"
                        }
                    }
                }
            }
        },
        "/argoworkflow/{id}": {
            "get": {
                "description": "Get a ArgoWorkflow according to their ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a ArgoWorkflow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ArgoWorkflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Get a ArgoWorkflow from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "delete a ArgoWorkflow with a specified id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ArgoWorkflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update a ArgoWorkflow from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "update a ArgoWorkflow with a specified id",
                "parameters": [
                    {
                        "description": "update ArgoWorkflow",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateArgoWorkflowRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "ArgoWorkflow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/argoworkflows": {
            "get": {
                "description": "List ArgoWorkflows according to the page_id and page_size",
                "produces": [
                    "application/json"
                ],
                "summary": "List all ArgoWorkflows",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/envlayer": {
            "post": {
                "description": "Create a EnvLayer and saves it in postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new EnvLayer",
                "parameters": [
                    {
                        "description": "Create EnvLayer",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateEnvLayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SwaggerEnvLayer"
                        }
                    }
                }
            }
        },
        "/envlayer/{id}": {
            "delete": {
                "description": "Get a EnvLayer from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "delete a EnvLayer with a specified id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "EnvLayer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "description": "update a EnvLayer from postgresql",
                "produces": [
                    "application/json"
                ],
                "summary": "update a EnvLayer with a specified id",
                "parameters": [
                    {
                        "description": "update EnvLayer",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateEnvLayerRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "EnvLayer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/envlayers": {
            "get": {
                "description": "List EnvLayers according to the page_id and page_size",
                "produces": [
                    "application/json"
                ],
                "summary": "List all EnvLayers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PageID",
                        "name": "page_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "api.CreateArgoWorkflowRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "api.CreateEnvLayerRequest": {
            "type": "object"
        },
...
        "api.SwaggerArgoWorkflow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "api.SwaggerEnvLayer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_request_id": {
                    "type": "integer"
                },
                "environment_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "process_id": {
                    "type": "integer"
                },
                "s3_path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdateArgoWorkflowRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "git_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.UpdateEnvLayerRequest": {
            "type": "object"
        },
...
    }
}